+++ src/components/TaskSummary.tsx
@@ -2,12 +2,11 @@ import { useEffect, useState } from "react";
 import { supaClient } from "../api/client";
 import TaskCard from "./TaskCard";
 import { Task } from "../api/global.types";
-import TaskForm from "./TaskForm";
 
 const TaskSummary = () => {
   const [tasks, setTasks] = useState<Task[] | null>([]);
   const [loading, setLoading] = useState<boolean>(true);
-  const [showForm, setShowForm] = useState<boolean>(false);
+
   const getData = async () => {
     const {
       data: { user },
@@ -18,7 +17,7 @@ const TaskSummary = () => {
         .select("*, pets!inner(*,users_pets!inner(*))")
         .eq("pets.users_pets.user_id", user.id)
         .neq("is_completed", true);
-        data?.sort((a,b)=>a.task_id-b.task_id)
+      data?.sort((a, b) => a.task_id - b.task_id);
       setTasks(data);
       setLoading(false);
     }
@@ -43,22 +42,8 @@ const TaskSummary = () => {
       .subscribe();
   }, []);
 
-  const popUpTaskForm = () => {
-    setShowForm(true);
-  };
-
   return (
     <div>
-      {showForm ? (
-        <TaskForm setShowForm={setShowForm} />
-      ) : (
-        <button
-          className="bg-lightblue px-20 py-2 rounded-full font-extrabold text-white hover:bg-mediumblue border-solid border-mediumblue border-b-4 border-r-2 hover:border-lightblue"
-          onClick={popUpTaskForm}
-        >
-          Add New Task
-        </button>
-      )}
       {loading ? (
         <p>loading tasks</p>
       ) : (
